% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cnvkit_data.R
\name{check_cnvkit_data}
\alias{check_cnvkit_data}
\title{This function check if .cnr and .cns files have the correct structure for the plots.}
\usage{
check_cnvkit_data(cnr_data, cns_data)
}
\arguments{
\item{cnr_data}{Dataframe containing .cnr file originated from the tool CNVkit.}

\item{cns_data}{Dataframe containing .cns file originated from the tool CNVkit.}
}
\value{
Nothing
}
\description{
This function check if .cnr and .cns files have the correct structure for the plots.
}
