% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_gene.R
\docType{data}
\name{gene}
\alias{gene}
\title{Plot gene log2 coverage}
\format{
An object of class \code{character} of length 1.
}
\usage{
gene
}
\arguments{
\item{cnr_data}{Dataframe containing .cnr file originated from the tool CNVkit.}

\item{cns_data}{Dataframe containing .cns file originated from the tool CNVkit.}

\item{genes}{A string with a gene name (HGNC ID). Default: "TP53"}

\item{all.transcripts}{Boolean value for make the plot showing all transcripts of the target gene. Default = FALSE}

\item{gene_structure}{Boolean value for make the plot showing canonical transcript of the target gene. Default = TRUE}

\item{regulatory.elements}{Boolean value for make the plot displaying regulatory elements from ... Default: FALSE}

\item{log2_line_col}{Color for the average log2 line. Default: "deepskyblue"}

\item{log2_threshold}{Numeric value to create a threshold line in the plot. Defaul: NULL}

\item{log2_threshold_color}{Color for the log_threshold line. Default: "red"}
}
\value{
A matrix of the infile
}
\description{
This function create a plot displaying log2 values in a specific chromosome. You can give 
to this function one or more genes in form of a vector and the function will plot them
base on their position and width in plot.This function#'can also plot a chromosome 
icon which is colored base on log2 values. It can have a color gradient which goes from red,
when there is a negative log2 value,to green when there is a positive log2 value.
}
\keyword{datasets}
