% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_chr.R
\name{plot_all_chrs}
\alias{plot_all_chrs}
\title{Plot all chromosome log2 coverages}
\usage{
plot_all_chrs(
  cnr_data,
  cns_data,
  only_autosomal = FALSE,
  chr_text_size = 4,
  chr_colors = NULL,
  log2_line_col = "red"
)
}
\arguments{
\item{cnr_data}{Dataframe containing .cnr file originated from the tool CNVkit.}

\item{cns_data}{Dataframe containing .cns file originated from the tool CNVkit.}

\item{only_autosomal}{Parameter to plot only autosomal chromosomes. Default: FALSE}

\item{chr_text_size}{Size of the chromosome number text displayed at the top of the plot. Default: 4
If you don't want any text set chr_text_size = NULL}

\item{chr_colors}{A vector of colors. Colors will be used om the plot  base on their order in the vector. I you plot only autosomes you must input 22 colors,
if you plot also sex chromosomes you must input 24 colors. Default: NULL}

\item{log2_line_col}{Color for the average log2 line. Default: "red"}
}
\value{
A ggplot graph.
}
\description{
This function create a plot displaying log2 ratios in all chromosomes.
}
