% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_chr.R
\name{plot_single_chr}
\alias{plot_single_chr}
\title{Plot single chromosome log2 coverages}
\usage{
plot_single_chr(
  cnr_data,
  cns_data,
  chr = "chr9",
  chr_picture = FALSE,
  genes = NULL,
  log2_line_col = "deepskyblue",
  log2_threshold = NULL,
  log2_threshold_color = "red"
)
}
\arguments{
\item{cnr_data}{Dataframe containing .cnr file originated from the tool CNVkit.}

\item{cns_data}{Dataframe containing .cns file originated from the tool CNVkit.}

\item{chr}{Chromosome the function should plot The accepted codes are:
"chr1","chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9", "chr10", "chr11",
"chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18", "chr19", "chr20", "chr21",
"chr22", "chrX", "chrY". Default: "chr1"}

\item{chr_picture}{Plot a cromosome icon colored base on log2 values. Default: FALSE}

\item{genes}{A string with a gene or a vector containing gene names (HGNC ID). Default: NULL}

\item{log2_line_col}{Color for the average log2 line. Default: "deepskyblue"}

\item{log2_threshold}{Numeric value to create a threshold line in the plot. Defaul: NULL}

\item{log2_threshold_color}{Color for the log_threshold line. Default: "red"}
}
\value{
A ggplot graph.
}
\description{
This function create a plot displaying log2 values in a specific chromosome. You can give 
to this function one or more genes in form of a vector and the function will plot them
base on their position and width in plot.This function#'can also plot a chromosome 
icon which is colored base on log2 values. It can have a color gradient which goes from red,
when there is a negative log2 value,to green when there is a positive log2 value.
}
